+++ src/base/PluginTrait.php
@@ -1,0 +1,68 @@
+<?php
+namespace verbb\kint\base;
+
+use verbb\kint\services\Service;
+
+use Craft;
+
+use yii\log\Logger;
+
+use verbb\base\BaseHelper;
+
+trait PluginTrait
+{
+    // Static Properties
+    // =========================================================================
+
+    public static $plugin;
+
+
+    // Public Methods
+    // =========================================================================
+
+    public static function log($message, $attributes = []): void
+    {
+        if ($attributes) {
+            $message = Craft::t('kint', $message, $attributes);
+        }
+
+        Craft::getLogger()->log($message, Logger::LEVEL_INFO, 'kint');
+    }
+
+    public static function error($message, $attributes = []): void
+    {
+        if ($attributes) {
+            $message = Craft::t('kint', $message, $attributes);
+        }
+
+        Craft::getLogger()->log($message, Logger::LEVEL_ERROR, 'kint');
+    }
+
+
+    // Public Methods
+    // =========================================================================
+
+    public function getService(): Service
+    {
+        return $this->get('service');
+    }
+
+
+    // Private Methods
+    // =========================================================================
+
+    private function _setPluginComponents(): void
+    {
+        $this->setComponents([
+            'service' => Service::class,
+        ]);
+
+        BaseHelper::registerModule();
+    }
+
+    private function _setLogging(): void
+    {
+        BaseHelper::setFileLogging('kint');
+    }
+
+}