+++ src/controllers/BaseController.php
@@ -1,0 +1,24 @@
+<?php
+namespace verbb\kint\controllers;
+
+use verbb\kint\Kint;
+
+use craft\web\Controller;
+
+use yii\web\Response;
+
+class BaseController extends Controller
+{
+    // Public Methods
+    // =========================================================================
+
+    public function actionSettings(): Response
+    {
+        $settings = Kint::$plugin->getSettings();
+
+        return $this->renderTemplate('kint/settings', [
+            'settings' => $settings,
+        ]);
+    }
+
+}