+++ CHANGELOG.md
@@ -1,9 +1,14 @@
-# Kint Changelog
+# Changelog
 
-All notable changes to this project will be documented in this file.
+## 2.0.0 - 2022-08-22
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
+> {note} The plugin’s package name has changed to `verbb/kint`. Autologin will need be updated to 2.0 from a terminal, by running `composer require verbb/kint && composer remove mildlygeeky/craft-kint`.
+
+### Changed
+- Migration to `verbb/kint`.
+- Now requires Craft 3.7+.
 
 ## 1.0.0 - 2019-02-05
+
 ### Added
 - Initial release